+++ search-mcp-proxy/main.ts
@@ -1,19 +1,28 @@
 import { withSimplerAuth } from "simplerauth-client";
-//Search MCP proxy server (https://search-mcp.parallel.ai) that proxies `/mcp` to https://mcp.parallel.ai/v1beta/search_mcp and adds minimally needed additions to make it work with oauth. CAREFULLY test and document these additions!
-
 export default {
   fetch: withSimplerAuth(
     async (request, env, ctx) => {
       const url = new URL(request.url);
       if (url.pathname === "/mcp") {
-        const req = new Request("https://mcp.parallel.ai/v1beta/search_mcp", {
-          body: request.body,
-          headers: request.headers,
-          method: request.method,
-        });
-        return fetch(req);
+        const headers = new Headers(request.headers);
+        headers.set(
+          "x-api-key",
+          request.headers.get("Authorization")?.slice("Bearer ".length)
+        );
+        headers.delete("Authorization");
+
+        const response = await fetch(
+          "https://mcp.parallel.ai/v1beta/search_mcp/",
+          {
+            body: request.body,
+            headers,
+            method: request.method,
+          }
+        );
+
+        return response;
       }
-      return new Response("Not found", { status: 404 });
+      return new Response("Not found!", { status: 404 });
     },
     {
       isLoginRequired: false,